feat(*): delete CrowdIn files if empty
