+++ dev/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "payload": "^1.10.4"
+    "payload": "^1.10.5"
   },
   "devDependencies": {
     "@types/express": "^4.17.17",
