+++ package.json
@@ -29,7 +29,7 @@
     "@types/deep-equal": "^1.0.1",
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.2",
-    "@types/node": "^20.3.2",
+    "@types/node": "^20.3.3",
     "babel-jest": "^29.5.0",
     "commitizen": "^4.3.0",
     "get-port": "^7.0.0",
@@ -51,12 +51,12 @@
     ]
   },
   "dependencies": {
-    "@crowdin/crowdin-api-client": "^1.23.2",
+    "@crowdin/crowdin-api-client": "^1.23.3",
     "@types/dot-object": "^2.1.2",
     "deep-equal": "^2.2.1",
     "dot-object": "^2.1.4",
     "lodash": "^4.17.21",
-    "payload": "^1.10.4",
+    "payload": "^1.10.5",
     "slate-serializers": "^0.4.0"
   },
   "volta": {
